Do not use the 16-bit 'short' type to keep track of numeric precision or scale when building the schema table for a query.  Fix for [ef2216192d].

+++ System.Data.SQLite/SQLiteDataReader.cs
@@ -866,8 +866,8 @@ namespace System.Data.SQLite
       tbl.Columns.Add(SchemaTableColumn.ColumnName, typeof(String));
       tbl.Columns.Add(SchemaTableColumn.ColumnOrdinal, typeof(int));
       tbl.Columns.Add(SchemaTableColumn.ColumnSize, typeof(int));
-      tbl.Columns.Add(SchemaTableColumn.NumericPrecision, typeof(short));
-      tbl.Columns.Add(SchemaTableColumn.NumericScale, typeof(short));
+      tbl.Columns.Add(SchemaTableColumn.NumericPrecision, typeof(int));
+      tbl.Columns.Add(SchemaTableColumn.NumericScale, typeof(int));
       tbl.Columns.Add(SchemaTableColumn.IsUnique, typeof(Boolean));
       tbl.Columns.Add(SchemaTableColumn.IsKey, typeof(Boolean));
       tbl.Columns.Add(SchemaTableOptionalColumn.BaseServerName, typeof(string));
